activate option for breaking at waterbodies
This does not break the test... if the organize_independent networks is not run with the new
breaking function, this commit will have no impact.

+++ src/python_framework_v02/troute/nhd_network.py
@@ -66,7 +66,7 @@ def extract_connections(rows, target_col, terminal_code=0):
 def extract_waterbodies(rows, target_col, waterbody_null=-9999):
     """Extract waterbody mapping from dataframe.
     """
-    return rows.loc[rows[target_col] != waterbody_null, target_col].to_dict()
+    return rows.loc[rows[target_col] != waterbody_null, target_col].astype("int").to_dict()
 
 
 def reverse_surjective_mapping(d):
